various cleanups
